spqr JWT security

+++ src/main/java/com/school/repository/UserRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
 
     public List<User> findByEnabledTrue();
 
     @Query("select u from User u where u.username like :username and u.id <> :userId ")
-    Optional<User> findByUsernameExept(String username, Integer userId);
+    Optional<User> findByUsernameExept(String username, Long userId);
 }
